pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "financial-offer-ranking-ml-poc"
version = "0.1.0"
description = "PoC: Personalized Credit Offer Propensity & Ranking using UCI Bank Marketing"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5.0",
    "scikit-learn>=1.2.0",
    "numpy>=1.23.0",
    "scipy>=1.9.0",
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "joblib>=1.2.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
